use mediator pattern

+++ Ecommerce_PhuongNam_v1.Application/Interfaces/Address/IDistrictService.cs
@@ -4,6 +4,6 @@ using Ecommerce_PhuongNam_v1.Application.Paging;
 
 namespace Ecommerce_PhuongNam_v1.Application.Interfaces.Address;
 
-public interface IDistrictService : IService<DistrictCreate, DistrictUpdate, int, DistrictResponse, PagingRequest, PagingResult<DistrictResponse>>
+public interface IDistrictService : IService<DistrictCreate, DistrictUpdate, long, DistrictResponse, PagingRequest, PagingResult<DistrictResponse>>
 {
 }
