+++ Ecommerce_PhuongNam_v1.Application/Interfaces/Address/IProvinceService.cs
@@ -3,7 +3,7 @@ using Ecommerce_PhuongNam_v1.Application.DTOs.Address.Responses.Province;
 
 namespace Ecommerce_PhuongNam_v1.Application.Interfaces.Address;
 
-public interface IProvinceService : IService<ProvinceCreate, ProvinceUpdate, int, ProvinceResponse, object, object>
+public interface IProvinceService : IService<ProvinceCreate, ProvinceUpdate, long, ProvinceResponse, object, object>
 {
     
 }
